GEN find templates automatically
