+++ templates/web/api/generate_server.sh.tpl
